# Simplify how finalizers are ran at shutdown