@peterzhu2118 2024-08-15
This requires a fix in Tempfile in https://github.com/ruby/tempfile/pull/31